project/requirements.txt
pyyaml
pydantic
email-validator
pytest
arrow
loguru
pycountry
jinja2
xmlschema
requests